reflected recent class separation to Tests as well
