+++ test/TestBasicNeeds.py
@@ -1,25 +1,10 @@
 import os
-from common.BasicNeeds import BasicNeeds
+from sources.common.BasicNeeds import BasicNeeds
 # package to facilitate multiple operation system operations
-import platform
 import unittest
 
 
-class MyTestCase(unittest.TestCase):
-
-    def test_file_statistics(self):
-        parent_folder = os.path.abspath(os.pardir) + '/' + '/'.join([
-            'virtual_environment',
-            'Scripts'
-        ])
-        relevant_file = 'pip'
-        if platform.system() == 'Windows':
-            relevant_file += '.exe'
-        file_to_evaluate = os.path.join(parent_folder, relevant_file)
-        bn = BasicNeeds()
-        value_to_assert = bn.fn_get_file_statistics(file_to_evaluate)['size [bytes]']
-        value_to_compare_with = os.path.getsize(file_to_evaluate)
-        self.assertEqual(value_to_assert, value_to_compare_with)
+class TestBasicNeeds(unittest.TestCase):
 
     def test_numbers_with_leading_zero(self):
         pair_values = [
